add selectionBehavior functionality
