+++ packages/main/src/components/AnalyticalTable/hooks/useTableRowStyling.ts
@@ -1,11 +1,16 @@
-import { enrichEventWithDetails } from '@ui5/webcomponents-react-base/lib/Utils';
 import { TableSelectionMode } from '@ui5/webcomponents-react/lib/TableSelectionMode';
+import { TableSelectionBehavior } from '@ui5/webcomponents-react/lib/TableSelectionBehavior';
+import { toggleSelected } from '../util';
 
 const ROW_SELECTION_ATTRIBUTE = 'data-is-selected';
 
 export const useTableRowStyling = (hooks) => {
   hooks.getRowProps.push((passedRowProps, { instance, row }) => {
-    const { classes, selectionMode, onRowSelected, isTreeTable } = instance.webComponentsReactProperties;
+    const {
+      classes,
+      selectionMode,
+      selectionBehavior
+    } = instance.webComponentsReactProperties;
     const isEmptyRow = row.original?.emptyRow;
     let className = classes.tr;
     if (row.isGrouped) {
@@ -22,47 +27,14 @@ export const useTableRowStyling = (hooks) => {
       role: 'row'
     };
     if ([TableSelectionMode.SINGLE_SELECT, TableSelectionMode.MULTI_SELECT].includes(selectionMode) && !isEmptyRow) {
-      rowProps.onClick = (e) => {
-        if (row.isGrouped) {
-          return;
-        }
-        if (isTreeTable) {
-          if (selectionMode === TableSelectionMode.MULTI_SELECT) {
-            instance.dispatch({
-              type: 'SET_SELECTED_ROWS',
-              selectedIds: Object.assign({}, ...instance.selectedFlatRows.map((item) => ({ [item.id]: true })), {
-                [row.id]: !row.isSelected
-              })
-            });
-          } else {
-            instance.dispatch({ type: 'SET_SELECTED_ROWS', selectedIds: { [row.id]: !row.isSelected } });
+      if ([TableSelectionBehavior.ROW_ONLY, TableSelectionBehavior.ROW].includes(selectionBehavior)) {
+        rowProps.onClick = (e) => {
+          if (row.isGrouped) {
+            return;
           }
-        } else {
-          row.toggleRowSelected();
-        }
-
-        if (typeof onRowSelected === 'function') {
-          const payload = {
-            row,
-            isSelected: !row.isSelected
-          };
-          const payloadWithFlatRows = {
-            ...payload,
-            selectedFlatRows: !row.isSelected
-              ? [...instance.selectedFlatRows, row]
-              : instance.selectedFlatRows.filter((prevRow) => prevRow.id !== row.id)
-          };
-          onRowSelected(
-            enrichEventWithDetails(e, TableSelectionMode.MULTI_SELECT === selectionMode ? payloadWithFlatRows : payload)
-          );
-        }
-
-        if (selectionMode === TableSelectionMode.SINGLE_SELECT && !isTreeTable) {
-          instance.selectedFlatRows.forEach(({ id }) => {
-            instance.toggleRowSelected(id, false);
-          });
-        }
-      };
+          toggleSelected(e, instance, row);
+        };
+      }
       if (row.isSelected) {
         rowProps[ROW_SELECTION_ATTRIBUTE] = '';
       }
