+++ packages/main/src/components/AnalyticalTable/util/index.ts
@@ -1,4 +1,6 @@
+import { enrichEventWithDetails } from '@ui5/webcomponents-react-base/lib/Utils';
 import { defaultOrderByFn } from 'react-table';
+import { TableSelectionMode } from '@ui5/webcomponents-react/lib/TableSelectionMode';
 
 export const orderByFn = (rows, functions, directions) => {
   const wrapSortFn = (sortFn, index) => {
@@ -21,4 +23,42 @@ export const orderByFn = (rows, functions, directions) => {
   const wrappedSortfunctions = functions.map(wrapSortFn);
 
   return defaultOrderByFn(rows, wrappedSortfunctions, directions);
+};
+
+export const toggleSelected = (e, row, instance) => {
+  const { isTreeTable, selectionMode, onRowSelected } = instance.webComponentsReactProperties;
+  if (isTreeTable) {
+    if (selectionMode === TableSelectionMode.MULTI_SELECT) {
+      instance.dispatch({
+        type: 'SET_SELECTED_ROWS',
+        selectedIds: Object.assign({}, ...instance.selectedFlatRows.map((item) => ({ [item.id]: true })), {
+          [row.id]: !row.isSelected
+        })
+      });
+    } else {
+      instance.dispatch({ type: 'SET_SELECTED_ROWS', selectedIds: { [row.id]: !row.isSelected } });
+    }
+  } else {
+    row.toggleRowSelected();
+  }
+  if (typeof onRowSelected === 'function') {
+    const payload = {
+      row,
+      isSelected: !row.isSelected
+    };
+    const payloadWithFlatRows = {
+      ...payload,
+      selectedFlatRows: !row.isSelected
+        ? [...instance.selectedFlatRows, row]
+        : instance.selectedFlatRows.filter((prevRow) => prevRow.id !== row.id)
+    };
+    onRowSelected(
+      enrichEventWithDetails(e, TableSelectionMode.MULTI_SELECT === selectionMode ? payloadWithFlatRows : payload)
+    );
+  }
+  if (selectionMode === TableSelectionMode.SINGLE_SELECT && !isTreeTable) {
+    instance.selectedFlatRows.forEach(({ id }) => {
+      instance.toggleRowSelected(id, false);
+    });
+  }
 };
